Initialize repository
This adds a base README with some bare-bones details of for our group.
We'll need to flesh it out later.

By putting this on GitHub, it'll allow us all to make PRs, and it will
also allow us to subscribe to group updates in a format that most of us
likely use already.